User_service/requirements.txt
boto3==1.12.15
botocore==1.15.15
cffi==1.14.0
Click==7.0
cryptography==2.8
dnspython==1.16.0
docutils==0.15.2
email-validator==1.0.5
Flask==1.1.1
Flask-Mail==0.9.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
Flask-Cors==3.0.8
pyotp==2.3.0
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
